requirements.txt
django
djangorestframework
django-allauth
django-phonenumber-field[phonenumbers]
drf-nested-routers
djangorestframework-simplejwt
django-filter
python-dotenv